move more to composer-errs
